// apps/rs-marketplace/types.ts
export interface Product {
  id: string;
  name: string;
  seller: string;
  price: number;
  costPerItem?: number;
  compareAtPrice?: number;
  currency: string;
  shortDescription: string;
  description: string;
  images: string[];
  rating: number;
  reviewCount: number;
  collectionId: string | null;
  status: 'Ativo' | 'Rascunho';
  inventory: number;
  type: string;
  sku?: string;
  barcode?: string;
  requiresShipping: boolean;
  trackQuantity: boolean;
  chargeTax: boolean;
  continueSelling: boolean;
  seoTitle?: string;
  seoDescription?: string;
  urlHandle?: string;
  options: ProductOption[];
  variants: ProductVariant[];
  weight?: number;
  weightUnit?: 'kg' | 'g' | 'lb' | 'oz';
  supplier?: string;
}

export interface ProductOption {
  id: string;
  name: string;
  values: string[];
}

export interface ProductVariant {
  id: string;
  options: { [key: string]: string };
  price: number;
  inventory: number;
  sku?: string;
  imageId?: string;
}
// ...
export interface Collection {
  id: string;
  title: string;
  description: string;
  imageUrl: string;
  productIds: string[];
}
// ...
export interface Review {
  id: string;
  productId: string;
  productName: string;
  author: string;
  rating: number;
  title: string;
  text: string;
  createdAt: string;
  status: 'Aprovada' | 'Pendente' | 'Rejeitada';
}

export interface Answer {
  id: string;
  author: string;
  text: string;
  createdAt: string;
}

export interface Question {
  id: string;
  productId: string;
  author: string;
  text: string;
  createdAt: string;
  answers: Answer[];
}